fix: refactor code, create movie-card and sorting-settings-container, replace img with Image, add token to useAuth
