+++ next.config.ts
@@ -1,7 +1,9 @@
 import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
-  /* config options here */
+  images: {
+    domains: ["fkvhqaugajnlkvvlapoz.supabase.co"],
+  },
 };
 
 export default nextConfig;
