+++ src/components/item-with-icon.tsx
